local-training/requirements.txt
tensorflow-macos
tensorflow-metal
boto3
tensorflowjs
